Use Case 1 Tri's update

+++ Source Code/Case Study 1/Market/Market/Models/Orders_Has_Products.cs
@@ -17,7 +17,7 @@ namespace Market.Models
         public int orderId { get; set; }
         public int productId { get; set; }
         public int quantity { get; set; }
-        public double pricePerUnit { get; set; }
+        public decimal pricePerUnit { get; set; }
     
         public virtual Order Order { get; set; }
         public virtual Product Product { get; set; }
